+++ src/App.tsx
@@ -8,7 +8,6 @@ import ForgetPassword from '@modules/forgot-password/ForgotPassword';
 import RecoverPassword from '@modules/recover-password/RecoverPassword';
 import { useWindowSize } from '@app/hooks/useWindowSize';
 import { calculateWindowSize } from '@app/utils/helpers';
-import { useDispatch, useSelector } from 'react-redux';
 import { setWindowSize } from '@app/store/reducers/ui';
 import ReactGA from 'react-ga4';
 
@@ -19,44 +18,39 @@ import Profile from '@pages/profile/Profile';
 
 import PublicRoute from './routes/PublicRoute';
 import PrivateRoute from './routes/PrivateRoute';
-import { setAuthentication } from './store/reducers/auth';
-import {
-  GoogleProvider,
-  getAuthStatus,
-  getFacebookLoginStatus,
-} from './utils/oidc-providers';
+import { setCurrentUser } from './store/reducers/auth';
+
+import { firebaseAuth } from './firebase';
+import { onAuthStateChanged } from 'firebase/auth';
+import { useAppDispatch, useAppSelector } from './store/store';
 
 const { VITE_NODE_ENV } = import.meta.env;
 
 const App = () => {
   const windowSize = useWindowSize();
-  const screenSize = useSelector((state: any) => state.ui.screenSize);
-  const dispatch = useDispatch();
+  const screenSize = useAppSelector((state) => state.ui.screenSize);
+  const dispatch = useAppDispatch();
   const location = useLocation();
 
   const [isAppLoading, setIsAppLoading] = useState(true);
 
-  const checkSession = async () => {
-    try {
-      let responses: any = await Promise.all([
-        getFacebookLoginStatus(),
-        GoogleProvider.getUser(),
-        getAuthStatus(),
-      ]);
-
-      responses = responses.filter((r: any) => Boolean(r));
-
-      if (responses && responses.length > 0) {
-        dispatch(setAuthentication(responses[0]));
-      }
-    } catch (error: any) {
-      console.log('error', error);
-    }
-    setIsAppLoading(false);
-  };
-
   useEffect(() => {
-    checkSession();
+    onAuthStateChanged(
+      firebaseAuth,
+      (user) => {
+        if (user) {
+          dispatch(setCurrentUser(user as any));
+        } else {
+          dispatch(setCurrentUser(undefined));
+        }
+        setIsAppLoading(false);
+      },
+      (e) => {
+        console.log(e);
+        dispatch(setCurrentUser(undefined));
+        setIsAppLoading(false);
+      }
+    );
   }, []);
 
   useEffect(() => {