+++ src/index.tsx
@@ -9,7 +9,7 @@ import * as serviceWorker from './serviceWorker';
 import { BrowserRouter } from 'react-router-dom';
 import ReactGA from 'react-ga4';
 
-const { VITE_NODE_ENV, VITE_GA_ID } = import.meta.env;
+export const { VITE_NODE_ENV, VITE_GA_ID } = import.meta.env;
 
 if (VITE_NODE_ENV === 'production' && VITE_GA_ID) {
   ReactGA.initialize(VITE_GA_ID);
@@ -17,6 +17,7 @@ if (VITE_NODE_ENV === 'production' && VITE_GA_ID) {
 
 const container: any = document.getElementById('root');
 const root = createRoot(container);
+
 root.render(
   <Provider store={store}>
     <BrowserRouter>