+++ src/pages/profile/TimelineTab.tsx
@@ -1,3 +1,4 @@
+import {} from '@app/index';
 import { Image } from '@profabric/react-components';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';