+++ src/routes/PrivateRoute.tsx
@@ -1,9 +1,9 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useSelector } from 'react-redux';
+import { useAppSelector } from '@app/store/store';
 
 const PrivateRoute = () => {
-  const isLoggedIn = useSelector((state: any) => state.auth.authentication);
-  return isLoggedIn ? <Outlet /> : <Navigate to="/login" />;
+  const isLoggedIn = useAppSelector((state) => state.auth.currentUser);
+  return isLoggedIn ? <Outlet /> : <Navigate to={`/login`} />;
 };
 
 export default PrivateRoute;