+++ src/routes/PublicRoute.tsx
@@ -1,9 +1,9 @@
+import { useAppSelector } from '@app/store/store';
 import { Navigate, Outlet } from 'react-router-dom';
-import { useSelector } from 'react-redux';
 
 const PublicRoute = () => {
-  const isLoggedIn = useSelector((state: any) => state.auth.authentication);
-  return isLoggedIn ? <Navigate to="/" /> : <Outlet />;
+  const isLoggedIn = useAppSelector((state) => state.auth.currentUser);
+  return isLoggedIn ? <Navigate to={`/`} /> : <Outlet />;
 };
 
 export default PublicRoute;