+++ src/store/store.ts
@@ -1,8 +1,9 @@
-import { configureStore } from '@reduxjs/toolkit';
+import { Action, ThunkAction, configureStore } from '@reduxjs/toolkit';
 
 import { authSlice } from '@app/store/reducers/auth';
 import { uiSlice } from '@app/store/reducers/ui';
 import { createLogger } from 'redux-logger';
+import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 
 const store = configureStore({
   reducer: {
@@ -10,7 +11,23 @@ const store = configureStore({
     ui: uiSlice.reducer,
   },
   middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(createLogger()) as any,
+    getDefaultMiddleware({ serializableCheck: false }).concat(
+      createLogger()
+    ) as any,
 });
 
 export default store;
+
+export const useAppDispatch = () => useDispatch<ReduxDispatch>();
+export const useAppSelector: TypedUseSelectorHook<ReduxState> = useSelector;
+
+/* Types */
+export type ReduxStore = typeof store;
+export type ReduxState = ReturnType<typeof store.getState>;
+export type ReduxDispatch = typeof store.dispatch;
+export type ReduxThunkAction<ReturnType = void> = ThunkAction<
+  ReturnType,
+  ReduxState,
+  unknown,
+  Action
+>;