+++ vite.config.ts
@@ -1,13 +1,16 @@
 import { defineConfig, loadEnv } from 'vite';
 import react from '@vitejs/plugin-react';
 import path from 'path';
+import 'dotenv/config';
 
 // eslint-disable-next-line import/no-anonymous-default-export
 export default ({ mode }) => {
   process.env = { ...process.env, ...loadEnv(mode, process.cwd()) };
 
+  const { VITE_NODE_ENV } = process.env;
+
   return defineConfig({
-    mode: 'development',
+    mode: VITE_NODE_ENV,
     plugins: [react()],
     resolve: {
       alias: {